- Moved the loading animation function in the code folder - Renamed "loading" to "loadingUtils" - Updated imports
